Changed Fit back to take into account length of sequences.

+++ Python/Fit.py
@@ -42,7 +42,7 @@ class Fit(object):
             ccCount = min_ol.ccCount()
             if di.number_crossings() < 3:
                 bonus = 10000
-                fitness = 1.0 + bonus/(di.number_crossings()**float(1) + ccCount**float(3) + ol.length()**float(0) + (ol.length() - min_ol.length())**float(2) + 1.0)
+                fitness = 1.0 + bonus/(di.number_crossings()**float(1) + ccCount**float(3) + ol.length()**float(1) + (ol.length() - min_ol.length())**float(2) + 1.0)
             else:
                 bonus = 1
                 fitness = 1.0 + bonus/(di.number_crossings()**float(2) + ccCount**float(1) + ol.length()**float(0) + (ol.length() - min_ol.length())**float(2) + 1.0)
